rename the `interpolations` module to `interpolation`
