+++ xarray_healpy/interpolation/mask.py
